requirements.txt
tensorflow-gpu==1.5.0
numpy==1.14.0
